fast-python-lang/pyproject.toml
[tool.poetry]
name = "fast-python-lang"
version = "0.1.0"
description = "string similarity performance comparison for fast-python"
authors = ["Nick Anthony <nanthony007@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.0.1"
pytest-benchmark = {extras = ["histogram"], version = "^3.4.1"}
fastDamerauLevenshtein = "^1.0.7"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
rich = "^12.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [
    "tests",
]
